Move the EntityAddInfo to re_viewport_blueprint to make it availalble to other crates
